pyproject.toml
[tool.poetry]
name = "blog-poc"
version = "0.1.0"
description = ""
authors = ["aorri2 <aorri96@ikw.kr>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "4.2"
djangorestframework = "3.15.2"
drf-spectacular = "0.27.2"
asgiref = "3.8.1"
attrs = "24.2.0"
inflection = "0.5.1"
jsonschema = "4.23.0"
jsonschema-specifications = "2024.10.1"
pyyaml = "6.0.2"
referencing = "0.35.1"
rpds-py = "0.21.0"
sqlparse = "0.5.2"
uritemplate = "4.1.1"
psycopg2-binary = "^2.9.10"
python-decouple = "^3.8"
django-redis = "^5.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120
indent-width = 4

[tool.ruff.lint]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]
select = [
    "F",
    "E",
    "W",
    "I",
    "N",
    "D",
    "UP",
    "YTT",
    "ANN",
    "ASYNC",
    "S",
    "BLE",
    "FBT",
    "B",
    "A",
    "COM",
    "C",
    "DTZ",
    "T10",
    "DJ",
    "EM",
    "EXE",
    "FA",
    "ISC",
    "ICN",
    "LOG",
    "G",
    "PIE",
    "T20",
    "PYI",
    "PT",
    "Q",
    "RSE",
    "RET",
    "SLF",
    "SLOT",
    "SIM",
    "TID",
    "TCH",
    "INT",
    "ARG",
    "PTH",
    "TD",
    "PD",
    "PGH",
    "PLC",
    "PLE",
    "PLR",
    "PLW",
    "TRY",
    "FLY",
    "NPY",
    "PERF",
    "FURB",
]

[tool.ruff.format]
quote-style = "single"